requirements.txt
requests
psycopg2-binary
sshtunnel
python-dotenv
streamlit
pandas
